Fix DemonImage imports by extracting component
